pokedexServer/requirements.txt
Pillow==11.0.0 # ImageOps, Image, ImageDraw, ImageFont
requests==2.32.3